@@ -13,6 +13,7 @@ import java.io.IOException;
 import java.util.LinkedList;
 
 import javax.imageio.ImageIO;
+import javax.swing.BoxLayout;
 import javax.swing.ImageIcon;
 import javax.swing.JFrame;
 import javax.swing.JLabel;
@@ -69,7 +70,7 @@ public class ConfigWindow {
 		try {
 			BufferedImage bufferedImage = ImageIO.read(getClass().getResourceAsStream("/icon.jpg"));
 			this.mainFrame.setIconImage(bufferedImage);
-		} catch (IOException e) {
+		} catch (@SuppressWarnings("unused") IOException e) {
 			this.log.error("No Icon Found!");
 		}
 
@@ -120,13 +121,21 @@ public class ConfigWindow {
 
 		} else if (this.playlistViewGen.isPlaylistViewEnabled()) {
 
+			this.mainPanel.setLayout(new BoxLayout(mainPanel, BoxLayout.Y_AXIS));
+
+			JPanel playlistViewP = new JPanel();
 			for (JLabel pviewlabel : this.playlistViewGen.loadPlaylistView()) {
-				this.mainPanel.add(pviewlabel);
+				playlistViewP.add(pviewlabel);
 			}
+			this.mainPanel.add(playlistViewP);
+
+			JPanel staticsP = new JPanel();
 
 			for (JLabel sactionlabel : this.playlistViewGen.loadStaticActionsView()) {
-				this.mainPanel.add(sactionlabel);
+				staticsP.add(sactionlabel);
 			}
+
+			this.mainPanel.add(staticsP);
 
 		} else {
 
@@ -168,7 +177,6 @@ public class ConfigWindow {
 		addSong.setText("Add Song");
 		addSong.addActionListener(new ActionListener() {
 
-			@SuppressWarnings("unused")
 			@Override
 			public void actionPerformed(ActionEvent e) {
 				new SongAddPanel();
@@ -179,7 +187,6 @@ public class ConfigWindow {
 		addLabel.setText("Add Label");
 		addLabel.addActionListener(new ActionListener() {
 
-			@SuppressWarnings("unused")
 			@Override
 			public void actionPerformed(ActionEvent e) {
 				new LabelAddPanel();
